Avoid alternatives on high-level rules (Union, SubType).
Leads to many backtracks, which is not efficient at all without
memoization enabled. Since memoization seems to have a bug that
makes it slower on some schemas, this strategy seems easier...

The parser AST is no longer a purely semantics tree, but a real
syntax tree, which is not necessarily a big problem in practice.
Indeed, the AST generated by syntax classes is kept the same through
fixing of `to_ast` rules in Union and SubType.

This change is 100% backward compatible, thereby.