Add support for typed extra attributes in MultiTupleType.
This commit adds support for typed extra attributes in multi-tuple
type syntax. E.g., the following type:

```
{ a: Integer, ...: Integer }
```

requires a tuple with at least one `a` attribute which is an Integer,
but also allows additional attributes, provided they are Integers too.
In the example, the following data is correct:

```
{ a: 12, b: 1827, c: 1876 }
```

but the following is not:

```
{ a: 12, b: "foobar" }
```

The implementation is similar to what has been done in finitio.js, and
relies on the `allow_extra` Heading option that now captures the type
expected for extra attributes. It defaults to AnyType (i.e. `.`).

Backward compatibility of the Heading class still allows specifying
{allow_extra : true}, which is the same as specifying AnyType itself
as extra type.