Implement subtype and ADT structural equivalence.
This commit ensures that sub types (resp. ADTs) are considered
structurally equivalent - according to the `==` method, as other
types - when having the same super type and constraints (resp.
the same base type and contracts).

This is implemented through a fake ProcWithCode class, that
overrides the `hash`, `==` and `eql?` method of the resulting
Proc of constraints/contracts expression compilation.